Set of costs added to the objective
FlowThinning_Esati.py - is the python code for FT formulation problem. Cost function it is added the the objective.
AffineFlowThinning_Esati.py - is the python code for AFT formulation problem. Cost function it is added the the objective.
example.xlsx - is the excel file that contains the input data for the example problem (from task 1.2). Cos function is added in an extra sheet.

Results_FlowThinning.xlsx - The decision variables are printed in this excel file for FT problem.
Results_AffineFlowThinning.xlsx  - The decision variables are printed in this excel file for AFT problem.