Print Constraints in Excel files
The new update contains the code for printing the constraints in Excel files with variables, with values of variables and calculation for the left hand side of the constraints. This helps on evaluation of the variable values for FT and AFT problem formulation and to compare the results with the herustics algorithms.

New results are found on Results_FlowThinning.xlsx excel file for flow thinning - Sheet name: Verification of Constraints.
New results are found on Results_AffineFlowThinning.xlsx excel file for affine flow thinning - Sheet name: Verification of Constraints.